requirements.txt
docopt==0.6.2
PyYAML==3.12
tabulate==0.7.5
python-digitalocean==1.9.0
